refactor: abstract social auth
